Fixed System.Windows.Markup.XamlParseException
Something goes wrong with the invocation of the constructor on type "VirtualBoxIntegration.MainWindow" with an invalid cast exception.

It looks like the error is caused when declaring _VirtualBox = VirtualBox.Virtualbox() and Including the VirtualBox header seems to fix the issues.

Also added an error debugging header for the exception window.